+++ App/Sagas/likedChannels/index.js
@@ -3,10 +3,11 @@ import Promise from 'bluebird';
 import { assign } from 'lodash';
 import { call, put, select } from 'redux-saga/effects';
 
-import type {TChannel, TChannelStore, APIResponse, TRootState, TLike} from '../types/';
-import {PER_PAGE} from '../constants';
-import {likedChannelsActions} from '../Redux/';
-import {likesRef, channelsRef, snapshotExists, statusCode, isSuccess, channelStoreArrayToActiveObject} from '../Services/';
+import type {TChannel, TChannelStore, APIResponse, TRootState, TLike} from '../../types/';
+import {PER_PAGE} from '../../constants';
+import {likedChannelsActions} from '../../Redux/';
+import {likesRef, channelsRef, snapshotExists, statusCode, isSuccess, channelStoreArrayToActiveObject, likesPostToFirebase} from '../../Services/'; // eslint-disable-line
+import {likedChannelsSelector, uidSelector} from '../selector'; // eslint-disable-line
 
 export const getStartAt = (state: TRootState) => state.likedChannels.startAt;
 
@@ -78,5 +79,17 @@ export function* getLikedChannels<T>(): Generator<T, any, any> {
   const channelsPromise: Array<Promise<TChannelStore>> = yield call(getChannels, responce.snapshot);
   const channelsArray: TChannelStore[] = yield call(Promise.all, channelsPromise);
   const channels: {[key: string]: TChannelStore} = channelStoreArrayToActiveObject(channelsArray);
+  // const localChennels: {[key: string]: TChannelStore} = yield select(likedChannelsSelector);
+  // const uid = yield select(uidSelector);
+  // Object.keys(localChennels).forEach((key) => {
+  //   if (!channels[key]) {
+  //     call(likesPostToFirebase.likesNew, key, localChennels[key].likeCount, uid);
+  //   }
+  //   const isDiff = channels[key] && localChennels[key].likeCount > channels[key].likeCount;
+  //   if (isDiff) {
+  //     const diff = localChennels[key].likeCount - channels[key].likeCount;
+  //     call(likesPostToFirebase.likes, key, diff, uid);
+  //   }
+  // });
   yield put(likedChannelsActions.likedChannelsSuccess(channels));
 }
